Uplift to Pages
The home page is set as the "Enter Queue" page for guests. Includes a login/logout page for staff and admin. First draft of the admin page has been created. The Customer page has been updated to include a (static) queue number and the menu. Staff page has been transitioned to use cards, and the image of the restaurant layout has been added.